I gang med type checker
